Move tests for PDAStack to an own file
